+++ tests/dummy/app/JsonApi/Posts/Validators.php
@@ -31,8 +31,8 @@ class Validators extends AbstractValidators
      */
     protected $allowedSortParameters = [
         'id',
-        'created-at',
-        'updated-at',
+        'createdAt',
+        'updatedAt',
         'title',
         'slug',
     ];
@@ -53,7 +53,7 @@ class Validators extends AbstractValidators
     protected $allowedIncludePaths = [
         'author',
         'comments',
-        'comments.created-by',
+        'comments.createdBy',
         'image',
         'tags',
     ];
@@ -99,9 +99,10 @@ class Validators extends AbstractValidators
 
     /**
      * @param Post|null $record
+     * @param array $data
      * @return array|mixed
      */
-    protected function rules($record = null): array
+    protected function rules($record, array $data): array
     {
         $slugUnique = 'unique:posts,slug';