+++ tests/dummy/app/JsonApi/Suppliers/Validators.php
@@ -31,14 +31,14 @@ class Validators extends AbstractValidators
      * @var array
      */
     protected $allowedSortParameters = [
-        'created-at',
-        'updated-at',
+        'createdAt',
+        'updatedAt',
     ];
 
     /**
      * @inheritDoc
      */
-    protected function rules($record = null): array
+    protected function rules($record, array $data): array
     {
         return [
             'name' => ['required', 'string'],