+++ tests/dummy/app/JsonApi/Users/Schema.php
@@ -44,10 +44,10 @@ class Schema extends SchemaProvider
     public function getAttributes($resource)
     {
         return [
-            'created-at' => $resource->created_at->toAtomString(),
+            'createdAt' => $resource->created_at,
             'email' => $resource->email,
             'name' => $resource->name,
-            'updated-at' => $resource->updated_at->toAtomString(),
+            'updatedAt' => $resource->updated_at,
         ];
     }