+++ tests/dummy/app/JsonApi/Users/Validators.php
@@ -27,8 +27,8 @@ class Validators extends AbstractValidators
      * @var array
      */
     protected $allowedSortParameters = [
-        'created-at',
-        'updated-at',
+        'createdAt',
+        'updatedAt',
         'name',
         'email',
     ];
@@ -57,7 +57,7 @@ class Validators extends AbstractValidators
     /**
      * @inheritDoc
      */
-    protected function rules($record = null): array
+    protected function rules($record, array $data): array
     {
         $rules = [
             'name' => 'required|string',
@@ -68,7 +68,7 @@ class Validators extends AbstractValidators
         ];
 
         if (!$record) {
-            $rules['password-confirmation'] = 'required_with:password|same:password';
+            $rules['passwordConfirmation'] = 'required_with:password|same:password';
         }
 
         return $rules;